backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["GoppyKrishna-77 <goppykrishna7077@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
urllib3 = "1.26.15"
flask = "^3.1.0"
hydra-core = "^1.3.2"
pydantic = {extras = ["email"], version = "^2.10.6"}
sqlalchemy = "^2.0.37"
db-sqlite3 = "^0.0.1"
redis = "^5.2.1"
python-dotenv = "^1.0.1"
appwrite = "^7.1.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
